requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
razorpay>=1.4,<2.0

gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
